Removed shared code from python http stager, now uses jinja includes
